Move racing sleep test to JVM
